+++ core/initializer/test_execution/test_executor/test_executor.go
@@ -98,7 +98,7 @@ Returns:
 */
 func RunTest(
 		executionInstanceId uuid.UUID,
-		ctx context.Context,
+		testSetupContext context.Context,
 		log *logrus.Logger,
 		dockerClient *client.Client,
 		suiteExecutionVolume string,
@@ -120,6 +120,12 @@ func RunTest(
 	}
 	log.Info("Docker manager created successfully")
 
+	// We'll use the test setup context for setting stuff up so that a cancellation (e.g. Ctrl-C)
+	//  will prevent any new things from getting added to Docker. We still want to be able to retrieve exit codes
+	//  and logs after a Ctrl-C though, so we use the background context for doing those tasks (rather than the
+	//  potentially-cancelled setup context).
+	testTeardownContext := context.Background()
+
 	log.Infof("Creating Docker network for test with subnet mask %v...", subnetMask)
 	freeIpAddrTracker, err := commons.NewFreeIpAddrTracker(
 		log,
@@ -137,7 +143,7 @@ func RunTest(
 		time.Now().Format(networkNameTimestampFormat),
 		executionInstanceId.String(),
 		testName)
-	networkId, err := dockerManager.CreateNetwork(ctx, networkName, subnetMask, gatewayIp)
+	networkId, err := dockerManager.CreateNetwork(testSetupContext, networkName, subnetMask, gatewayIp)
 	if err != nil {
 		// TODO If the user Ctrl-C's while the CreateNetwork call is ongoing then the CreateNetwork will error saying
 		//  that the Context was cancelled as expected, but *the Docker engine will still create the networks!!! We'll
@@ -145,7 +151,7 @@ func RunTest(
 		//  networks with our network name and delete them
 		return false, stacktrace.Propagate(err, "Error occurred creating Docker network %v for test %v", networkName, testName)
 	}
-	defer removeNetworkDeferredFunc(log, dockerManager, networkId)
+	defer removeNetworkDeferredFunc(testTeardownContext, log, dockerManager, networkId)
 	log.Infof("Docker network %v created successfully", networkId)
 
 	// TODO use hostnames rather than IPs, which makes things nicer and which we'll need for Docker swarm support
@@ -180,7 +186,7 @@ func RunTest(
 
 	log.Info("Launching testsuite container to run the test...")
 	testRunningContainerId, err := testsuiteLauncher.LaunchTestRunningContainer(
-		ctx,
+		testSetupContext,
 		dockerManager,
 		networkId,
 		suiteExecutionVolume,
@@ -217,7 +223,7 @@ func RunTest(
 		testRunningContainerIp,
 		suiteExecutionVolume)
 	kurtosisApiContainerId, err := dockerManager.CreateAndStartContainer(
-		ctx,
+		testSetupContext,
 		kurtosisApiImageName,
 		networkId,
 		kurtosisApiIp,
@@ -243,7 +249,7 @@ func RunTest(
 	// The Kurtosis API will be our indication of whether the test suite container stopped within the timeout or not
 	log.Info("Waiting for Kurtosis API container to exit...")
 	kurtosisApiExitCode, err := dockerManager.WaitForExit(
-		context.Background(),
+		testTeardownContext,
 		kurtosisApiContainerId)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred waiting for the exit of the Kurtosis API container: %v", err)
@@ -252,7 +258,7 @@ func RunTest(
 	// At this point, we may be printing the logs of a stopped test suite container, or we may be printing the logs of
 	//  still-running container that's exceeded the hard test timeout. Regardless, we want to print these so the user
 	//  gets more information about what's going on, and the user will learn the exact error below
-	banner_printer.PrintContainerLogsWithBanners(*dockerManager, ctx, testRunningContainerId, log, testRunningContainerDescription)
+	banner_printer.PrintContainerLogsWithBanners(*dockerManager, testTeardownContext, testRunningContainerId, log, testRunningContainerDescription)
 
 	var testStatusRetrievalError error
 	switch kurtosisApiExitCode {
@@ -291,7 +297,7 @@ func RunTest(
 
 	// The test suite container will already have stopped, so now we get the exit code
 	testSuiteExitCode, err := dockerManager.WaitForExit(
-		context.Background(),
+		testTeardownContext,
 		testRunningContainerId)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred retrieving the test suite container exit code")
@@ -337,11 +343,13 @@ func buildApiContainerEnvVarsMap(
 Helper function for making a best-effort attempt at removing a network and the containers inside after a test has
 	exited (either normally or with error)
 */
-func removeNetworkDeferredFunc(log *logrus.Logger, dockerManager *docker_manager.DockerManager, networkId string) {
+func removeNetworkDeferredFunc(
+		testTeardownContext context.Context,
+		log *logrus.Logger,
+		dockerManager *docker_manager.DockerManager,
+		networkId string) {
 	log.Infof("Attempting to remove Docker network with id %v...", networkId)
-	// We use the background context here because we want to try and tear down the network even if the context the test was running in
-	//  was cancelled. This might not be right - the right way to do it might be to pipe a separate context for the network teardown to here!
-	if err := dockerManager.RemoveNetwork(context.Background(), networkId, networkTeardownContainerStopTimeout); err != nil {
+	if err := dockerManager.RemoveNetwork(testTeardownContext, networkId, networkTeardownContainerStopTimeout); err != nil {
 		log.Errorf("An error occurred removing Docker network with ID %v:", networkId)
 		log.Error(err.Error())
 		log.Error("NOTE: This means you will need to clean up the Docker network manually!!")